release: v0.5.2 (#91)
This is Morio v0.5.2 🎉 

This includes a fix for OIDC which in itself was the reason to make this a patch release.
In addition, we have removed the dedicated `/preseed` and `/validate/preseed` endpoints from the API.

They were historically added because `/setup` expects a full settings object, whereas `/preseed` only expects the preseed settings. So, since they required different validation against the data schema, they were split.
However, when only validating the preseeding settings, there is no guardrails against any issues in the settings sourced from the preseeing settings. 

So, instead we now (temporarily) resolve the preseeded settings (if a base settings file is configured) and then validate the entire (preseeded) settings against the data schema.

This makes for a more meaningful validation step, and it also means there is no longer a reason to have different `/preseed` and `/setup` endpoints as `/setup` now handles both scenarios.

The changelog for this release is:

**Fixed**

- [api] Fix incorrect key property when creating JWT in OIDC identity provider (d2b53829defd1d887a208610e24cc955d6ec604f)
- [core] Do not assume preseed.base is a string, handle objects too

**Removed**

- [api] We no longer have a dedicated `/preseed` endpoint, use the `/setup` endpoint instead
- [api] We no longer have a dedicated `/validate/preseed` endpoint, use the `/validate/settings` endpoint instead
- [core] We no longer have a dedicated `/preseed` endpoint, use the `/setup` endpoint instead
- [ui] Removed preseed upload UI