chore: Fix API unit tests, avoid unit test specifics (#157)
This fixes https://github.com/certeu/morio/issues/104

This also contains more changes, which are all with the goal of avoiding
doing 'special' things when running unit tests. In a couple of places in
the codebase and config, we were taking a different branch in the code
when detecting a unit test because -- at the time -- we didn't have a
way to make the unit tests function properly without them.

So, this has been refactored a bit to remove this. In the process of
doing so, the following changes were made as part of this commit:

**Changed**

- We removed all configuration that was handled differently for unit
  tests
- [api] Remove anything done differently for unit tests
- [api] Persist test credentials to disk so tests can be re-run
  incrementally
- [core] Remove anything done differently for unit tests
- [core] Persist test credentials to disk so tests can be re-run
  incrementally

**Fixed**

- [api] RBAC would block access when a token in a cookie was expired,
  even if a valid Bearer token was provided
- [api] Update test run-script with prefixed container name
- [api] Required length for `api_key` and `api_key_secret` were inverted
  in the schema
- [core] Rolled back the change for
  [https://github.com/certeu/morio/issues/138](https://github.com/certeu/morio/issues/138) as the fix did
  not address the root issue.

---------

Co-authored-by: serge-tellene <114212556+serge-tellene@users.noreply.github.com>