chore: Properly implement support for different release channels (#168)
We have three release channels now: stable, canary, and testing. 
This makes the required changes to ensure that we can build container
images and packages for each channel, as well as making sure that when
deploying Morio, it actually uses all images from that channel.

Our APT repositories were specifically tied to Debian 12 (bookwork).
However, there's not reason for this, as they will work just the same on
a different Debian version, or Ubuntu or another distro using apt.

So we have updated our infrastructure to ensure our repositories are
version agnostic.

In addition, this includes changes to how the packages are build that
make them more discoverable for new contributors. Specifically, apt
source files are not included in the repo, rather than generated
dynamically with a script.
---------

Co-authored-by: serge-tellene <114212556+serge-tellene@users.noreply.github.com>